% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{base64_encode_raw}
\alias{base64_encode_raw}
\title{Encode a raw vector as base64}
\usage{
base64_encode_raw(r)
}
\arguments{
\item{r}{A raw vector.}
}
\value{
A single character string, padded with \code{=} as the standard requires.
}
\description{
Encode a raw vector as base64
}
\keyword{internal}
