% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{catch_all_label}
\alias{catch_all_label}
\title{The label assigned to cells inside the parent gate matching no population}
\usage{
catch_all_label()
}
\value{
A single string.
}
\description{
The label assigned to cells inside the parent gate matching no population
}
\keyword{internal}
