% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{clin_is_numeric}
\alias{clin_is_numeric}
\title{Is a vector usable as a numeric clinical variable?}
\usage{
clin_is_numeric(v)
}
\arguments{
\item{v}{vector.}
}
\description{
Is a vector usable as a numeric clinical variable?
}
\keyword{internal}
