% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{default_value_map}
\alias{default_value_map}
\title{Default value translations; extend in the YAML config without touching code}
\usage{
default_value_map()
}
\description{
Default value translations; extend in the YAML config without touching code
}
\keyword{internal}
