% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{html_escape}
\alias{html_escape}
\title{Escape text for inclusion in HTML}
\usage{
html_escape(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Escape text for inclusion in HTML
}
\keyword{internal}
