% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{json_table}
\alias{json_table}
\title{Render a data.frame as a JSON array-of-arrays with a header}
\usage{
json_table(d)
}
\arguments{
\item{d}{a data.frame}
}
\description{
Render a data.frame as a JSON array-of-arrays with a header
}
\keyword{internal}
