% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miflowcyt.R
\name{kw_get}
\alias{kw_get}
\title{Pull one keyword out of an FCS keyword list, trying several spellings}
\usage{
kw_get(kw, ...)
}
\arguments{
\item{kw}{keyword list from \code{\link[=read_fcs_resolved]{read_fcs_resolved()}}}

\item{...}{candidate keyword names, in preference order}
}
\value{
the first non-empty value as a length-1 character, or NA
}
\description{
Pull one keyword out of an FCS keyword list, trying several spellings
}
\keyword{internal}
