% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-session.R
\name{load_session}
\alias{load_session}
\title{Restore a saved session
Usage:  st <- load_session("results/session_state.RData")
fig_umap_overview(st$embeddings$panel_1$cells, "new_plot.png")}
\usage{
load_session(path)
}
\arguments{
\item{path}{File path.}
}
\description{
Restore a saved session
Usage:  st <- load_session("results/session_state.RData")
fig_umap_overview(st$embeddings$panel_1$cells, "new_plot.png")
}
\keyword{internal}
