% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miflowcyt.R
\name{md_table}
\alias{md_table}
\title{Render a data.frame as a GitHub-flavoured markdown table}
\usage{
md_table(d)
}
\arguments{
\item{d}{a data.frame}
}
\description{
Render a data.frame as a GitHub-flavoured markdown table
}
\keyword{internal}
