% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-abundance.R
\name{p_annotation}
\alias{p_annotation}
\title{Format a p-value the way these figures conventionally annotate them}
\usage{
p_annotation(p, style = c("stars_and_trend", "stars", "numeric"))
}
\arguments{
\item{p}{The p.}

\item{style}{The style. Default \code{c("stars_and_trend", "stars", "numeric")}.}
}
\description{
Format a p-value the way these figures conventionally annotate them
}
\keyword{internal}
