% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{pretty_label}
\alias{pretty_label}
\title{Turn a snake_case column name into a readable axis/panel label}
\usage{
pretty_label(x)
}
\arguments{
\item{x}{A vector of values.}
}
\description{
Turn a snake_case column name into a readable axis/panel label
}
\keyword{internal}
