% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-by.R
\name{samples_at_timepoint}
\alias{samples_at_timepoint}
\title{The sample ids drawn at one timepoint}
\usage{
samples_at_timepoint(smap, level)
}
\arguments{
\item{smap}{Sample map.}

\item{level}{One timepoint level.}
}
\value{
Character vector of sample ids.
}
\description{
The sample ids drawn at one timepoint
}
\keyword{internal}
