% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-by.R
\name{split_dir_name}
\alias{split_dir_name}
\title{A filesystem-safe directory name for a split level}
\usage{
split_dir_name(x)
}
\arguments{
\item{x}{Character.}
}
\description{
A filesystem-safe directory name for a split level
}
\keyword{internal}
