% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{umap_discrete}
\alias{umap_discrete}
\title{UMAP coloured by a discrete variable}
\usage{
umap_discrete(
  df,
  colour_by,
  title = NULL,
  subtitle = NULL,
  point_size = NULL,
  alpha = NULL,
  legend_rows = NULL,
  palette = NULL,
  colors = fcs_colors()
)
}
\arguments{
\item{df}{The df.}

\item{colour_by}{The colour by.}

\item{title}{The title.}

\item{subtitle}{The subtitle.}

\item{point_size}{The point size.}

\item{alpha}{Significance threshold.}

\item{legend_rows}{The legend rows.}

\item{palette}{The palette.}

\item{colors}{Named list of colours; defaults to the package palette. See \code{\link[=fcs_colors]{fcs_colors()}}. Default \code{fcs_colors()}.}
}
\description{
UMAP coloured by a discrete variable
}
\keyword{internal}
