% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{write_sample_map_template}
\alias{write_sample_map_template}
\title{Write a sample-map template the user can fill in}
\usage{
write_sample_map_template(fcs_files, path, sample_ids = NULL)
}
\arguments{
\item{fcs_files}{The fcs files.}

\item{path}{File path.}

\item{sample_ids}{The sample ids.}
}
\description{
Write a sample-map template the user can fill in
}
\keyword{internal}
