% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-gates.R
\name{xml_escape}
\alias{xml_escape}
\title{Escape text for inclusion in XML}
\usage{
xml_escape(x)
}
\arguments{
\item{x}{character vector}
}
\value{
\code{x} with the five predefined entities replaced
}
\description{
Escape text for inclusion in XML
}
\keyword{internal}
